feat: implement non-custodial test for token association (CCS-188)
