+++ sample.env
@@ -3,7 +3,6 @@ OPENAI_API_KEY=
 # Operator
 HEDERA_ACCOUNT_ID=
 HEDERA_PRIVATE_KEY=
-HEDERA_PUBLIC_KEY=
 HEDERA_NETWORK=testnet
 HEDERA_KEY_TYPE=        # 'ECDSA' or 'ED25519'
