+++ src/tests/custodial/claim_airdrop.test.ts
@@ -1,11 +1,11 @@
 import { describe, beforeAll, expect, it, afterAll } from "vitest";
-import { AccountData } from "./utils/testnetUtils";
-import { LangchainAgent } from "./utils/langchainAgent";
-import { NetworkClientWrapper } from "./utils/testnetClient";
+import { AccountData } from "../utils/testnetUtils";
+import { LangchainAgent } from "../utils/langchainAgent";
+import { NetworkClientWrapper } from "../utils/testnetClient";
 import * as dotenv from "dotenv";
-import { HederaMirrorNodeClient } from "./utils/hederaMirrorNodeClient";
-import { NetworkType } from "./types";
-import { wait } from "./utils/utils";
+import { HederaMirrorNodeClient } from "../utils/hederaMirrorNodeClient";
+import { NetworkType } from "../types";
+import { wait } from "../utils/utils";
 
 const IS_CUSTODIAL = true;
