+++ src/tests/custodial/get_all_balances.test.ts
@@ -1,11 +1,11 @@
-import { HederaMirrorNodeClient } from "./utils/hederaMirrorNodeClient";
+import { HederaMirrorNodeClient } from "../utils/hederaMirrorNodeClient";
 import { describe, expect, it, beforeAll } from "vitest";
 import * as dotenv from "dotenv";
-import { NetworkClientWrapper } from "./utils/testnetClient";
-import { AccountData } from "./utils/testnetUtils";
-import { DetailedTokenBalance } from "./types";
-import { LangchainAgent } from "./utils/langchainAgent";
-import { wait } from "./utils/utils";
+import { NetworkClientWrapper } from "../utils/testnetClient";
+import { AccountData } from "../utils/testnetUtils";
+import { DetailedTokenBalance } from "../types";
+import { LangchainAgent } from "../utils/langchainAgent";
+import { wait } from "../utils/utils";
 
 const IS_CUSTODIAL = true;
