+++ src/tests/custodial/mint_nft.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, it, beforeAll, beforeEach } from "vitest";
-import { NetworkType} from "./types";
+import { NetworkType} from "../types";
 import * as dotenv from "dotenv";
-import { NetworkClientWrapper } from "./utils/testnetClient";
-import { HederaMirrorNodeClient } from "./utils/hederaMirrorNodeClient";
-import { LangchainAgent } from "./utils/langchainAgent";
-import { wait } from "./utils/utils";
+import { NetworkClientWrapper } from "../utils/testnetClient";
+import { HederaMirrorNodeClient } from "../utils/hederaMirrorNodeClient";
+import { LangchainAgent } from "../utils/langchainAgent";
+import { wait } from "../utils/utils";
 
 dotenv.config();
