+++ src/tests/custodial/transfer_hbar.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, it, beforeAll } from "vitest";
-import { LangchainAgent } from "./utils/langchainAgent";
-import { HederaMirrorNodeClient } from "./utils/hederaMirrorNodeClient";
+import { LangchainAgent } from "../utils/langchainAgent";
+import { HederaMirrorNodeClient } from "../utils/hederaMirrorNodeClient";
 import * as dotenv from "dotenv";
-import { NetworkClientWrapper } from "./utils/testnetClient";
-import { AccountData } from "./utils/testnetUtils";
-import { wait } from "./utils/utils";
+import { NetworkClientWrapper } from "../utils/testnetClient";
+import { AccountData } from "../utils/testnetUtils";
+import { wait } from "../utils/utils";
 
 const IS_CUSTODIAL = true;
